requirements.txt
fastapi
uvicorn
pandas
tiktoken
graphrag
pydantic
python-dotenv
asyncio
aiohttp
numpy
scikit-learn
matplotlib
seaborn
nltk
spacy
transformers
torch
torchvision
torchaudio
promptflow
llama-index
wget
llama-index-vector-stores-azureaisearch
llama-index-embeddings-azure-openai
llama-index-llms-azure-openai
azure-search-documents==11.6.0b3
azure-identity
langchain
langchain-openai